% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\name{anova.lc50}
\alias{anova.lc50}
\alias{anova.lc50list}
\title{Analysis of Deviance for lc50 model fits}
\usage{
\method{anova}{lc50}(object, ..., test = NULL)

\method{anova}{lc50list}(object, ..., test = NULL)
}
\arguments{
\item{object}{an object of class \code{lc50}, usually obtained as the
results from a call to \code{\link{lc50}}}

\item{...}{additional objects of class \code{lc50}.}

\item{test}{a character string, partially matching one of
"\code{LRT}", "\code{Chisq}", or "\code{Cp}". See
\code{link{stat.anova}}.}
}
\value{
An object of class \code{anova} inheriting from class
\code{data.frame}.
}
\description{
Compute an analysis of deviance table for an LC50 model fit.
}
\details{
Specifying a single object gives a sequential analysis of deviance
table for that fit. That is, the reductions in the residual
deviance as each term of the formula is added in turn are given in
as the rows of a table, plus the residual deviances themselves.

If more than one object is specified, the table has a row for the
residual degrees of freedom and deviance for each model. For all
but the first model, the change in degrees of freedom and deviance
is also given. (This only makes statistical sense if the models
are nested.) It is conventional to list the models from smallest
to largest, but this is up to the user.

When \code{test} is "LRT" or "Chisq" the table will contain test
statistics (and P values) comparing the reduction in deviance for
the row to the residuals.
}

