% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\docType{package}
\name{LC50-package}
\alias{LC50-package}
\title{Solar/Satellite Geolocation for Animal Tracking}
\description{
provides facilities for estimating lethal concentrations for a
toxin from destructively sampled survival data in the presence of
additional stressors and non-ignorable control mortality.
}
\author{
S. Wotherspoon, A. Proctor.
}

