% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\name{lc50.initialize}
\alias{lc50.initialize}
\title{Starting parameters for an LC50 model fit}
\usage{
lc50.initialize(Y, conc, group, link = c("probit", "logit"))
}
\arguments{
\item{Y}{a two column matrix of the number of survivals and mortalities in each sample.}

\item{conc}{a vector of tixin concentrations}

\item{group}{a factor delineating treatment groups}

\item{link}{the link function for survival fractions}
}
\value{
Return a list of with components
\item{\code{alpha}}{the rate parameter for each treatment group}
\item{\code{gamma}}{the probit of the control surival for each treatment group}
\item{\code{loglc50}}{the log lc50 for each treatment group}
}
\description{
Estimate starting parameters for and LC50 model fit
}
\details{
This is the default method for computing the starting values used
to initialize an \code{\link{lc50}} model.
}

