% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\name{lc50}
\alias{lc50}
\alias{lc50.fit}
\title{Estimate LC50 for a toxin}
\usage{
lc50(formula, concentration, group, data, start = NULL, link = c("probit",
  "logit"), common.background = FALSE, rate.shrink = 0,
  optim.control = list())

lc50.fit(X, Y, conc, group, alpha, beta, gamma, link, common.background,
  rate.shrink, optim.control = list())
}
\arguments{
\item{formula}{a formula relating log LC50 to covariates
describing the aditional stressors.}

\item{concentration}{the name of variable that is the
concentration of the toxin.}

\item{group}{a factor distinguishing treatment groups for the additional stressors.}

\item{data}{data frame containing variables in the model.}

\item{start}{Starting values used to initialize the model.  If
\code{start=NULL} these parameters are determined by
\code{\link{lc50.initialize}}.}

\item{link}{the link function for survival fractions}

\item{common.background}{should a common background survival be
estimated for each treatment group.}

\item{rate.shrink}{the shrinkage penalty for the rate parameters}

\item{optim.control}{control parameters for \code{optim}}

\item{X}{a design matrix}

\item{Y}{a two column matrix of responses}

\item{conc}{a vector of toxin concentrations}

\item{alpha}{vector of starting rate parameters}

\item{beta}{vector of starting coefficients}

\item{gamma}{vector of background survival parameters}
}
\value{
\code{lc50} returns an object of class inheriting from
"lc50". See later in this section.

The function \code{\link{summary}} (i.e.,
\code{link{summary.lc50}}) can be used to obtain or print a
summary of the results and the function \code{\link{anova}} (i.e.,
\code{\link{anova.lc50}}) to produce an analysis of deviance table
for the tests of additional stressor effects.

An LC50 model has several sets of coefficients, the generic
accessor function \code{\link{coef}} returns only the beta
coeffients.

An object of class "lc50" is a list containing at least the
following components:

\item{\code{logLik}}{the maximized log likelihood.}
\item{\code{aic}}{Akaike's information criteria.}
\item{\code{alpha}}{a vector of rate coefficients.}
\item{\code{alpha.cov}}{covariance of the rate coefficients.}
\item{\code{beta}}{a named vector of lc50 model coefficients.}
\item{\code{beta.cov}}{covariance of the lc50 model coefficients.}
\item{\code{gamma}}{a vector of background survival coefficients.}
\item{\code{gamma.cov}}{covariance of the background survival coefficients.}
\item{\code{coefficients}}{a named vector of lc50 model coefficients.}
\item{\code{cov.scaled}}{covariance of the lc50 model coefficients.}
\item{\code{loglc50}}{a named vector of log lc50s for the treatment groups.}
\item{\code{loglc50.cov}}{covariance of the lc50s for the treatment groups.}
\item{\code{concentration}}{a vector of taxin concentrations.}
\item{\code{group}}{a factor distinguishing treatment groups.}
\item{\code{x}}{a design matrix relating log lc50 to factors describing the additional stressors.}
\item{\code{y}}{a two column matrix of responses, giving the survivals and mortalities.}
\item{\code{fitted.values}}{the fitted probability of survival.}
\item{\code{residuals}}{the deviance residuals for the fit.}
\item{\code{deviance}}{the deviance for the fit.}
\item{\code{df.residual}}{the residual degrees of freedom.}
\item{\code{null.deviance}}{the deviance of the null model, which fits a single mortality rate to all data.}
\item{\code{df.null}}{the degrees of freedom for the null model.}
\item{\code{optim}}{the result of the call to \code{optim}.}
\item{\code{xlevels}}{a record of the levels of the factors used in fitting.}
\item{\code{contrasts}}{the contrasts used.}
\item{\code{call}}{the matched call.}
\item{\code{link}}{the link function.}
\item{\code{terms}}{the terms object used.}
\item{\code{model}}{the model frame.}
}
\description{
Estimate LC50 from survival data in the presence of additional
stressors and non-ignorable control mortality
}
\details{
DESCRIBE the model.

\code{lc50.fit} is the workhorse function: it is not normally
called directly but can be more efficient where the response
vector, design matrix and family have already been calculated.
}

