% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\name{lc50JAGS}
\alias{lc50JAGS}
\title{Estimate LC50 for a toxin}
\usage{
lc50JAGS(formula, concentration, group, data, start = NULL,
  link = c("probit", "logit"), common.background = FALSE, n.adapt = 500,
  n.chains = 4, alpha.mu = 0, alpha.tau = 0.001, beta.mu = 0,
  beta.tau = 0.001, gamma.mu = 0, gamma.tau = 0.001)
}
\arguments{
\item{formula}{a formula relating log LC50 to covariates
describing the aditional stressors.}

\item{concentration}{the name of variable that is the
concentration of the toxin.}

\item{group}{a factor distinguishing treatment groups for the additional stressors.}

\item{data}{data frame containing variables in the model.}

\item{start}{Starting values used to initialize the model.  If
\code{start=NULL} these parameters are determined by
\code{\link{lc50.initialize}}.}

\item{link}{the link function for survival fractions}

\item{common.background}{should a common background survival be
estimated for each treatment group.}

\item{n.adapt}{parameter passed to \code{jags.model}}

\item{n.chains}{parameter passed to \code{jags.model}}

\item{alpha.mu}{prior mean for alpha}

\item{alpha.tau}{prior precision for alpha}

\item{beta.mu}{either a single prior mean for all beta parameters,
or a vector of prior means, one for each parameter.}

\item{beta.tau}{either a single prior precision for all beta
parameters, or a vector of prior precisions, one for each
parameter.}

\item{gamma.mu}{prior mean for gamma}

\item{gamma.tau}{prior precision for gamma}
}
\value{
Returns an object inheriting from class \code{jags} which
can be used to generate dependent samples from the posterior
distribution of the parameters
}
\description{
Bayesian estimates of LC50 from survival data in the presence of additional
stressors and non-ignorable control mortality
}
\details{
This function is an analog of \code{\link{lc50}} that produces an
object of class \code{jags} which can be used to draw samples from
the posterior using \code{update} and \code{coda.samples} from
\pkg{rjags}.

The model assumes half Normal priors for \code{alpha} and Normal
priors for \code{beta} and \code{gamma}.  For \code{alpha} and
\code{gamma}, a single prior mean and precision is assumed for all
groups, for \code{beta} individual prior means and precisions can be specified
}

