% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC50.r
\name{lcx.initialize}
\alias{lcx.initialize}
\title{Starting parameters for an LCx model fit}
\usage{
lcx.initialize(Y, conc, group, link = c("probit", "logit"), lethal)
}
\arguments{
\item{Y}{a two column matrix of the number of survivals and mortalities in each sample.}

\item{conc}{a vector of tixin concentrations}

\item{group}{a factor delineating treatment groups}

\item{link}{the link function for survival fractions}

\item{lethal}{the modelled level of lethality}
}
\value{
Return a list of with components
\item{\code{alpha}}{the rate parameter for each treatment group}
\item{\code{gamma}}{the probit of the control surival for each treatment group}
\item{\code{loglcx}}{the log lcx for each treatment group}
}
\description{
Estimate starting parameters for and LCx model fit
}
\details{
This is the default method for computing the starting values used
to initialize an \code{\link{lcx}} model.
}

