% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC50.r
\name{lcxJAGS}
\alias{lcxJAGS}
\title{Estimate LCx for a toxin}
\usage{
lcxJAGS(formula, concentration, group, data, start = NULL,
  link = c("probit", "logit"), lethal = 50, common.background = FALSE,
  n.adapt = 500, n.chains = 4, alpha.mu = 0, alpha.tau = 1e-04,
  beta.mu = 0, beta.tau = 1e-04, gamma.mu = 0, gamma.tau = 1e-04)
}
\arguments{
\item{formula}{A formula relating log LCx to covariates describing
the additional stressors.}

\item{concentration}{The name of variable that is the
concentration of the toxin.}

\item{group}{A factor distinguishing treatment groups for the
additional stressors.}

\item{data}{Dataframe containing the variables in the model.}

\item{start}{Starting values used to initialize the model.  If
\code{start=NULL} these parameters are determined by
\code{\link{lcx.initialize}}.}

\item{link}{The link function for survival fractions.}

\item{lethal}{The level of lethality (ie "x") to be estimated.}

\item{common.background}{Should a common background survival be
estimated for each treatment group?}

\item{n.adapt}{Parameter passed to \code{jags.model}.}

\item{n.chains}{Parameter passed to \code{jags.model}.}

\item{alpha.mu}{Prior mean for alpha.}

\item{alpha.tau}{Prior precision for alpha.}

\item{beta.mu}{Either a single prior mean for all beta parameters,
or a vector of prior means, one for each parameter.}

\item{beta.tau}{Either a single prior precision for all beta
parameters, or a vector of prior precisions, one for each
parameter.}

\item{gamma.mu}{Prior mean for gamma.}

\item{gamma.tau}{Prior precision for gamma.}
}
\value{
Returns an object inheriting from class \code{jags} which
  can be used to generate dependent samples from the posterior
  distribution of the parameters
}
\description{
Bayesian estimates of LCx from survival data in the presence of additional
stressors and non-ignorable control mortality
}
\details{
This function is an analog of \code{\link{lcx}} that produces an
object of class \code{jags} which can be used to draw samples from
the posterior using \code{update} and \code{coda.samples} from
\pkg{rjags}.

The model assumes half Normal priors for \code{alpha} and Normal
priors for \code{beta} and \code{gamma}.  For \code{alpha} and
\code{gamma}, a single prior mean and precision is assumed for all
groups, for \code{beta} individual prior means and precisions can
be specified.
}
\seealso{
\code{\link{lcx}}
}

