% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\name{summary.lc50}
\alias{print.summary.lc50}
\alias{summary.lc50}
\title{Summmarizing LC50 model fits}
\usage{
\method{summary}{lc50}(object, background = TRUE, rate = FALSE, ...)

\method{print}{summary.lc50}(x, digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{an object of class \code{lc50}, obtained as the
result of a call to \code{\link{lc50}}}

\item{background}{if \code{TRUE} a summary table for the background survival is calculated}

\item{rate}{if \code{TRUE} a summary table for the rate parameters is calculated}

\item{...}{additional parameters are ignored.}

\item{x}{an object of class \code{summary.lc50}, usually, a result
of a call to \code{summary.lc50}.}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If \code{TRUE}, 'significance stars'
are printed for each coefficient.}
}
\value{
Returns an object of class \code{summary.lc50}, with components
\item{\code{coefficients}}{a table of coefficients.}
\item{\code{lc50}}{a table of LC50 for each treatment group.}
\item{\code{bsurv}}{a table of background survival for each treatment group.}
}
\description{
Summary method for class "\code{lc50}".
}

