% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC50.r
\name{summary.lcx}
\alias{print.summary.lcx}
\alias{summary.lcx}
\title{Summarizing LCx model fits}
\usage{
\method{summary}{lcx}(object, background = TRUE, rate = FALSE, ...)

\method{print}{summary.lcx}(x, digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{An object of class \code{lcx}, obtained as the
result of a call to \code{\link{lcx}}.}

\item{background}{If \code{TRUE} a summary table for the background survival is calculated.}

\item{rate}{If \code{TRUE} a summary table for the rate parameters is calculated,}

\item{...}{additional parameters are ignored.}

\item{x}{An object of class \code{summary.lcx}, usually, a result
of a call to \code{summary.lcx}.}

\item{digits}{The number of significant digits to use when printing.}

\item{signif.stars}{logical. If \code{TRUE}, 'significance stars'
are printed for each coefficient.}
}
\value{
Returns an object of class \code{summary.lcx}, with components
\item{\code{coefficients}}{a table of coefficients.}
\item{\code{lcx}}{a table of LCx for each treatment group.}
\item{\code{bsurv}}{optionally, a table of background survival for each treatment group.}
\item{\code{rate}}{optionally, a table of rates for each treatment group.}
}
\description{
Summary method for class "\code{lcx}".
}

