% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\docType{package}
\name{LC50-package}
\alias{LC50-package}
\title{LC50 lethal concentrations in the presence of additional stressors.}
\description{
The common 'critical effect concentration' used in toxicology bioassays is the Lethal concentration of 50% of the testing populaiton (LC50).
This package provides the facilities for estimating LC50s for a
toxin from sampled survival data at a single time-point in the presence of
additional stressors and non-ignorable control mortality. The result of which is the simultaneous calculations of multiple LC50s within a single experiemental framework where toxicants are tested within a framework of a factorial experiment.
Examples of facotrial arragement of multiple stressors include temperature, salinity,
}
\author{
S. Wotherspoon, A. Proctor.
}

