% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\name{lc50Jags}
\alias{lc50Jags}
\title{Estimate LC50 for a toxin}
\usage{
lc50Jags(formula, concentration, group, data, start = NULL,
  link = c("probit", "logit"), common.background = FALSE, n.adapt = 500,
  n.chains = 4)
}
\arguments{
\item{formula}{a formula relating log LC50 to covariates
describing the aditional stressors.}

\item{concentration}{the name of variable that is the
concentration of the toxin.}

\item{group}{a factor distinguishing treatment groups for the additional stressors.}

\item{data}{data frame containing variables in the model.}

\item{start}{Starting values used to initialize the model.  If
\code{start=NULL} these parameters are determined by
\code{\link{lc50.initialize}}.}

\item{link}{the link function for survival fractions}

\item{n.adapt}{parameter passed to \code{jags.model}}

\item{n.chains}{parameter passed to \code{jags.model}}
}
\value{
Returns an object inheriting from class \code{jags} which
can be used to generate dependent samples from the posterior
distribution of the parameters
}
\description{
Estimate LC50 from survival data in the presence of additional
stressors and non-ignorable control mortality
}
\details{
DESCRIBE the model.

\code{lc50.fit} is the workhorse function: it is not normally
called directly but can be more efficient where the response
vector, design matrix and family have already been calculated.
}

