% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC50.r
\name{lcx}
\alias{lcx}
\alias{lcx.fit}
\title{Estimate LCx for a toxin}
\usage{
lcx(formula, concentration, group, data, start = NULL, link = c("probit",
  "logit"), lethal = 50, quasi = FALSE, common.background = FALSE,
  rate.shrink = 0, optim.control = list())

lcx.fit(X, Y, conc, group, alpha, beta, gamma, link, lethal, quasi = FALSE,
  common.background = FALSE, rate.shrink = 0, optim.control = list())
}
\arguments{
\item{formula}{A formula relating log LCx to covariates
describing the additional stressors.}

\item{concentration}{The name of variable that is the
concentration of the toxin.}

\item{group}{A factor distinguishing treatment groups for the additional stressors.}

\item{data}{Dataframe containing the variables in the model.}

\item{start}{Starting values used to initialize the model.  If
\code{start=NULL} these parameters are determined by
\code{\link{lcx.initialize}}.}

\item{link}{The link function for survival fractions.}

\item{lethal}{The level of lethality (ie "x") to be estimated.}

\item{quasi}{Should a quasibinomial model be fitted?}

\item{common.background}{Should a common background survival be
estimated for each treatment group?}

\item{rate.shrink}{The shrinkage penalty for the rate parameters.}

\item{optim.control}{Control parameters for \code{optim}.}

\item{X}{A design matrix.}

\item{Y}{A two column matrix of responses.}

\item{conc}{A vector of toxin concentrations.}

\item{alpha}{The vector of starting rate parameters.}

\item{beta}{The vector of starting coefficients.}

\item{gamma}{The vector of background survival parameters.}
}
\value{
\code{lcx} returns an object of class inheriting from
"lcx". See later in this section.

The function \code{\link{summary}} (i.e.,
\code{link{summary.lcx}}) can be used to obtain or print a
summary of the results and the function \code{\link{anova}} (i.e.,
\code{\link{anova.lcx}}) to produce an analysis of deviance table
for the tests of additional stressor effects.

An LCx model has several sets of coefficients, the generic
accessor function \code{\link{coef}} returns only the beta
coefficients.

An object of class "lcx" is a list containing at least the
following components:

\item{\code{logLik}}{the maximized log likelihood.}
\item{\code{aic}}{Akaike's information criteria.}
\item{\code{alpha}}{a vector of rate coefficients.}
\item{\code{alpha.cov}}{covariance of the rate coefficients.}
\item{\code{beta}}{a named vector of lcx model coefficients.}
\item{\code{beta.cov}}{covariance of the lcx model coefficients.}
\item{\code{gamma}}{a vector of background survival coefficients.}
\item{\code{gamma.cov}}{covariance of the background survival coefficients.}
\item{\code{coefficients}}{a named vector of lcx model coefficients.}
\item{\code{cov.unscaled}}{covariance of the lcx model coefficients.}
\item{\code{loglcx}}{a named vector of log lcxs for the treatment groups.}
\item{\code{loglcx.cov}}{covariance of the lcxs for the treatment groups.}
\item{\code{concentration}}{a vector of taxin concentrations.}
\item{\code{group}}{a factor distinguishing treatment groups.}
\item{\code{x}}{a design matrix relating log lcx to factors describing the additional stressors.}
\item{\code{y}}{a two column matrix of responses, giving the survivals and mortalities.}
\item{\code{fitted.values}}{the fitted probability of survival.}
\item{\code{residuals}}{the deviance residuals for the fit.}
\item{\code{deviance}}{the deviance for the fit.}
\item{\code{df.residual}}{the residual degrees of freedom.}
\item{\code{dispersion}}{the dispersion.}
\item{\code{null.deviance}}{the deviance of the null model, which fits a single mortality rate to all data.}
\item{\code{df.null}}{the degrees of freedom for the null model.}
\item{\code{optim}}{the result of the call to \code{optim}.}
\item{\code{link}}{the link function.}
\item{\code{lethal}}{the modelled level of lethality.}
\item{\code{quasi}}{is the dispersion estimated.}
\item{\code{common.background}}{is background mortality common.}
\item{\code{rate.shrink}}{the shrinkage penalty.}
\item{\code{xlevels}}{a record of the levels of the factors used in fitting.}
\item{\code{contrasts}}{the contrasts used.}
\item{\code{call}}{the matched call.}
\item{\code{terms}}{the terms object used.}
\item{\code{model}}{the model frame.}
}
\description{
Estimate LCx from survival data in the presence of additional
stressors and non-ignorable control mortality.
}
\details{
The details of the model are described in the package vignette.

\code{lcx.fit} is the workhorse function: it is not normally
called directly but can be more efficient when the response
vector, design matrix and family have already been calculated.
}
\seealso{
\code{\link{lcxJAGS}}
}

