% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC50.r
\name{predict.lcx}
\alias{predict.lcx}
\title{Predict method for LCx fits}
\usage{
\method{predict}{lcx}(object, newdata, type = c("response", "adjusted",
  "link"), ...)
}
\arguments{
\item{object}{An object of class \code{lcx}, obtained as the
result of a call to \code{\link{lcx}}}

\item{newdata}{Optionally, a data frame in which to look for
variables with which to predict. If omitted, the fitted linear
predictors are used.}

\item{type}{The type of prediction required. If
\code{type="response"} predictions include both the background
mortality and the mortality due to the toxin, but if
\code{type="adjusted"} predictions only reflect mortality due to
the toxin.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Return a vector of predicted survival fractions.
}
\description{
Predict survival rates from a fitted LCx object.
}
\details{
If \code{newdata} is omitted the predictions are based on the data
used for the fit.  For \code{type="adjusted"}, it is assumed there
is no background mortality and predictions are made based purely
on the mortality due to the toxin. For \code{type="link"},
predictions are returned on the scale of the link function, again
ignoring any background mortality.
}
