% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LC50.r
\docType{data}
\name{toxicity}
\alias{toxicity}
\title{Simmulated toxicity data}
\format{A data frame with 225 rows and 8 variables:
\describe{
  \item{vial}{the replicate}
  \item{temperature}{temperature of solution}
  \item{salinity}{salinity of solution}
  \item{group}{additinal stressor treatment groups}
  \item{conc}{concentration of toxin in solution}
  \item{total}{number of individuals in vial}
  \item{alive}{number of survivors after 4 days}
  \item{dead}{number of dead individuals after 4 days}
}}
\usage{
toxicity
}
\description{
A simulated dataset showing individual survival following exposure
to a known toxin in the presence of the additional stressors
temperature and salinity.
}
\details{
Samples of an aqueous solution of a toxin, of varying
concentrations, are prepared for each of three salinities and are
held at one of three constant temperatures.  Approximately 10
individuals were added to each sample and the number of survivors
recorded at the end of four days exposure.
}
\keyword{datasets}

